% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_significance_csv}
\alias{export_significance_csv}
\title{Export significance results}
\usage{
export_significance_csv(out_dir, significance = NULL)
}
\arguments{
\item{out_dir}{Output directory.}

\item{significance}{A significance result object or named list of them.}
}
\value{
Export log tibble.
}
\description{
Export significance results
}
