% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_table_csv}
\alias{export_table_csv}
\title{Export a single table to CSV}
\usage{
export_table_csv(tbl, path, na = "")
}
\arguments{
\item{tbl}{Table to write.}

\item{path}{Destination path.}

\item{na}{String to use for missing values.}
}
\value{
Path invisibly.
}
\description{
Export a single table to CSV
}
