% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_standardize.R
\name{log_exclusions_init}
\alias{log_exclusions_init}
\title{Initialize an exclusion log}
\usage{
log_exclusions_init()
}
\value{
Empty tibble with standard exclusion-log columns.
}
\description{
Initialize an exclusion log
}
