% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_harmonization_table_merged}
\alias{plot_harmonization_table_merged}
\title{Plot chronology harmonization table}
\usage{
plot_harmonization_table_merged(chronology, save_path = NULL, base_size = 11)
}
\arguments{
\item{chronology}{Prepared chronology table or object containing it.}

\item{save_path}{Optional file path for saving.}

\item{base_size}{Base text size.}
}
\value{
ggplot object.
}
\description{
Plot chronology harmonization table
}
