% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_uncertainty_components}
\alias{plot_uncertainty_components}
\title{Plot uncertainty components}
\usage{
plot_uncertainty_components(components, save_path = NULL)
}
\arguments{
\item{components}{Output from summarize_uncertainty_components(), or tibble.}

\item{save_path}{Optional file path for saving.}
}
\value{
ggplot object.
}
\description{
Plot uncertainty components
}
