% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{test_peak_significance_global}
\alias{test_peak_significance_global}
\title{Test peak significance for the supra-regional/global curve}
\usage{
test_peak_significance_global(
  site_bucket_replicates,
  target_buckets,
  value_col = "value_norm",
  mode = c("neighbors", "baseline"),
  baseline_buckets = NULL,
  B = 1000L,
  seed = NULL
)
}
\description{
Test peak significance for the supra-regional/global curve
}
