% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{test_peak_significance_region}
\alias{test_peak_significance_region}
\title{Test peak significance for one regional curve}
\usage{
test_peak_significance_region(
  site_bucket_replicates,
  region_id,
  target_buckets,
  value_col = "value_norm",
  mode = c("neighbors", "baseline"),
  baseline_buckets = NULL,
  region_col = "site_region",
  B = 1000L,
  seed = NULL
)
}
\description{
Test peak significance for one regional curve
}
