% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{test_peak_significance_site}
\alias{test_peak_significance_site}
\title{Test peak significance for one site curve}
\usage{
test_peak_significance_site(
  site_bucket_replicates,
  site_id,
  target_buckets,
  value_col = "value_norm",
  mode = c("neighbors", "baseline"),
  baseline_buckets = NULL,
  B = 1000L,
  seed = NULL
)
}
\description{
Test peak significance for one site curve
}
